app/requirements.txt
shiny>=0.5.1
numpy>=1.26.0
clear-cache>=1.0.0
netCDF4>=1.6.4
PyYAML>=6.0.1
scikit-learn>=1.3.1
matplotlib>=3.8.0
torch>=2.0.1